wip: move some branch protector functions to shared file
